refactor: coding convention에 맞게 svg 파일을 snake_case로 변경
