+++ breaking-front/src/components/Header/Header.js
@@ -7,9 +7,9 @@ import Line from 'components/Line/Line';
 import Toggle from 'components/Toggle/Toggle';
 import ProfileImage from 'components/ProfileImage/ProfileImage';
 import * as Style from 'components/Header/Header.styles';
-import { ReactComponent as LogoIcon } from 'assets/svg/small-logo.svg';
+import { ReactComponent as LogoIcon } from 'assets/svg/small_logo.svg';
 import { ReactComponent as SearchIcon } from 'assets/svg/search.svg';
-import { ReactComponent as DropDownIcon } from 'assets/svg/drop-down.svg';
+import { ReactComponent as DropDownIcon } from 'assets/svg/drop_down.svg';
 import { ReactComponent as MoneyIcon } from 'assets/svg/money.svg';
 import { ReactComponent as SettingIcon } from 'assets/svg/setting.svg';
 import { ReactComponent as MyPageIcon } from 'assets/svg/mypage.svg';
