+++ breaking-front/src/pages/Post/Post.js
@@ -18,8 +18,8 @@ import { ReactComponent as CommentIcon } from 'assets/svg/comment.svg';
 import { ReactComponent as ETCIcon } from 'assets/svg/etc.svg';
 import { ReactComponent as EditIcon } from 'assets/svg/edit.svg';
 import { ReactComponent as RemoveIcon } from 'assets/svg/remove.svg';
-import { ReactComponent as BookmarkIcon } from 'assets/svg/small-bookmark.svg';
-import { ReactComponent as BookmarkedIcon } from 'assets/svg/small-bookmarked.svg';
+import { ReactComponent as BookmarkIcon } from 'assets/svg/small_bookmark.svg';
+import { ReactComponent as BookmarkedIcon } from 'assets/svg/small_bookmarked.svg';
 import { ReactComponent as ShareIcon } from 'assets/svg/share.svg';
 
 const Post = () => {
